+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2025-10-11
+
+### Changed
+- Updated `hightower-wireguard` dependency to version 0.2.1
+- Updated imports from `hightower_wireguard::transport` to `hightower_wireguard::connection` to match new API structure
+
 ## [0.1.3] - 2025-10-09
 
 ### Changed